+++ embc-app/ClientApp/src/app/evacuee-list/evacuee-list.component.html
@@ -1,3 +1,20 @@
-<p>
-  evacuee-list works!
-</p>
+<div class="mb-4">
+  <h1>ESS Team Dashboard</h1>
+  <h3>Welcome to your Dashboard!</h3>
+  <p>As an Emergency Support Services user you'll be able to register new evacuees, complete an existing evacuee
+    self- registration, or if needed, search and edit evacuee information.</p>
+  <h3 class="blue-header">LIST OF EVACUEES</h3>
+
+  <app-search-bar #searchBar (valueChange)="doSearch($event)" placeholder="Search by task number / last name / registration number / community ...">
+  </app-search-bar>
+
+  <!-- TODO: <a class="small" href="">advanced search</a> -->
+  <!-- TODO: show icons for HOH, RESTRICTED FILE -->
+
+  <!-- The start of the table stuff. -->
+  <!-- <p>1 - 10 of 2000 <span class="font-weight-bold">ACTIVE</span> evacuees
+          <small class="ml-3">(<a href="">show deactivated</a>)</small>
+        </p> -->
+  <app-evacuee-search-results #searchResultsView [searchResults]="searchResults$ | async">
+  </app-evacuee-search-results>
+</div>
