+++ embc-app/ClientApp/src/app/volunteer-list/volunteer-list.component.html
@@ -1,17 +1,58 @@
 <div class="mb-4">
+  <h1>ESS Team Management Dashboard</h1>
+  <p>Use this section to add new ESS users, as well as edit existing user information and permissions.</p>
+  <h3 class="blue-header">USERS LIST</h3>
 
-  <h1>ESS Team Dashboard</h1>
-  <h3>Welcome to your Dashboard!</h3>
-  <p>As an Emergency Support Services user you'll be able to register new evacuees, complete an existing evacuee
-    self- registration, or if needed, search and edit evacuee information.</p>
-  <h3 class="blue-header">LIST OF EVACUEES</h3>
+  <!-- <app-search-bar #searchBar (valueChange)="doSearch($event)" placeholder="Search for user by last name &#8230;">
+        </app-search-bar> -->
 
-  <app-search-bar #searchBar (valueChange)="doSearch($event)" placeholder="Search by task number / last name / registration number / community ...">
-  </app-search-bar>
-  <!-- The start of the table stuff. -->
-  <!-- <p>1 - 10 of 2000 <span class="font-weight-bold">ACTIVE</span> evacuees
-          <small class="ml-3">(<a href="">show deactivated</a>)</small>
-        </p> -->
-  <app-evacuee-search-results #searchResultsView [searchResults]="searchResults$ | async">
-  </app-evacuee-search-results>
+  <div role="search" class="form-row">
+    <div class="col-md-9">
+      <input #searchbox type="text" aria-label="search" class="form-control mb-2" placeholder="Search for user by last name &hellip;" (keydown.enter)="search(searchbox.value)">
+    </div>
+    <div class="col-md-3">
+      <button class="btn btn-primary" (click)="search(searchbox.value)">Search</button>
+    </div>
+  </div>
+  <div class="search-results mt-4">
+    <div *ngIf="metaVolunteers && metaVolunteers.data.length>0; then searchResultsTable; else notFound"></div>
+  </div>
+
+  <ng-template #searchResultsTable>
+    <!-- <p> {{metaVolunteers.metadata.page*metaVolunteers.metadata.pageSize}} - {{metaVolunteers.metadata.pageSize}} of
+            {{metaVolunteers.metadata.totalCount}} <span class="font-weight-bold">ACTIVE</span>
+            volunteers
+            <small class="ml-3">(<a href="">show deactivated</a>)</small>
+          </p> -->
+    <table class="table">
+      <thead class="thead-light">
+        <th scope="col">LAST NAME</th>
+        <th scope="col">FIRST NAME</th>
+        <th scope="col">BCEID</th>
+        <th scope="col">RESTRICTED FILES</th>
+        <th scope="col">ACTION</th>
+      </thead>
+      <tbody>
+        <tr *ngFor="let volunteer of metaVolunteers.data">
+          <td>{{volunteer.lastName}}</td>
+          <td>{{volunteer.firstName}}</td>
+          <td>{{volunteer.bceidAccountNumber}}</td>
+          <td>
+            <span *ngIf="volunteer.canAccessRestrictedFiles">Yes</span>
+            <span *ngIf="!volunteer.canAccessRestrictedFiles">No</span>
+          </td>
+          <td>
+            <a [routerLink]="" (click)="routeTo(volunteer.id)">Edit</a>
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </ng-template>
+
+  <ng-template #notFound>
+    <p>
+      {{notFoundMessage}}
+    </p>
+  </ng-template>
+
 </div>
