+++ embc-app/ClientApp/src/app/volunteer-list/volunteer-list.component.ts
@@ -1,11 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { SearchQueryParameters, EvacueeSearchResults } from '../shared/components/search';
-import { User, ListResult, Registration } from '../core/models';
-import { Observable } from 'rxjs';
-import { RegistrationService } from '../core/services/registration.service';
-import { Router } from '@angular/router';
-import { AuthService } from '../core/services/auth.service';
-import { map } from 'rxjs/operators';
+import { ListResult, Volunteer } from '../core/models';
+
+import { Router, ActivatedRoute } from '@angular/router';
+
+import { VolunteerService } from '../core/services/volunteer.service';
 
 @Component({
   selector: 'app-volunteer-list',
@@ -13,65 +11,35 @@ import { map } from 'rxjs/operators';
   styleUrls: ['./volunteer-list.component.scss']
 })
 export class VolunteerListComponent implements OnInit {
-
-
-  currentUser: User;
-  isLoggedIn = false;
-
-  // server response
-  resultsAndPagination$: Observable<ListResult<Registration>>;
-
-  // search related
-  isLoadingResults = false;
-  searchState = { offset: 0, limit: 100, sort: '', query: '' };
-  searchResults$: Observable<EvacueeSearchResults>;
-
+  // simple server response
+  metaVolunteers: ListResult<Volunteer>;
+  notFoundMessage: string = '';
   constructor(
-    private registrationService: RegistrationService,
+    private volunteerService: VolunteerService,
     private router: Router,
-    private authService: AuthService
+    private route: ActivatedRoute,
   ) { }
 
   ngOnInit() {
-    this.authService.getCurrentUser().subscribe(user => {
-      this.currentUser = user;
-      this.isLoggedIn = !!user;
-    });
-
-    // go get the data
-    this.doSearch();
+    // collect all volunteers
+    this.getVolunteers();
   }
 
-  doSearch(query: string = '') {
-    // update form state
-    this.searchState.query = query || '';
-    this.isLoadingResults = true;
-
-    // go get a fresh list of registrations from the service
-    const queryParams: SearchQueryParameters = {
-      offset: this.searchState.offset,
-      limit: this.searchState.limit,
-      sort: this.searchState.sort || '',
-      q: this.searchState.query
-    };
-    this.resultsAndPagination$ = this.registrationService.getRegistrations(queryParams);
-
-    // process server response into something we can display in the UI
-    this.searchResults$ = this.resultsAndPagination$.pipe(
-      map(x => {
-        // Flip flag to show that loading has finished.
-        this.isLoadingResults = false;
-        return { results: x.data, query } as EvacueeSearchResults;
-      })
-    );
-
-    // TODO: store the pagination metadata + links somewhere
-  }
-
-
-  routeTo(essFileNumber: string) {
+  routeTo(bceidAccountNumber: string) {
     // TODO: this seems like bad practive but fix when we have time
-    this.router.navigate(['register-evacuee/fill/' + essFileNumber]);
+    this.router.navigate(['../volunteer-edit/fill/' + bceidAccountNumber], { relativeTo: this.route });
   }
 
+  getVolunteers(limit?: number, offset?: number, query?: string, sort?: string) {
+    // get volunteers with supplied params defaults defined in
+    this.volunteerService.getVolunteers(limit, offset, query, sort).subscribe((v: ListResult<Volunteer>) => {
+      // save the metaVolunteers
+      this.metaVolunteers = v;
+    });
+  }
+
+  search(searchTerm: string) {
+    // submit and collect search
+    this.getVolunteers(null, null, searchTerm);
+  }
 }
